fixed authorization for admin

+++ routes/web.php
@@ -18,13 +18,13 @@ Auth::routes();
 
 Route::get('/', 'HomeController@index')->name('home');
 
-Route::prefix('admin')->middleware(['auth', 'can:isAdmin,App\User'])->group(function () {
+Route::prefix('admin')->middleware(['auth'])->group(function () {
 
 	//	USER
-	Route::get('users', 'UserController@index')->name('user.index');
-	Route::delete('users/{id}', 'UserController@destroy')->name('user.destroy');
-	Route::get('users/{id}/edit', 'UserController@edit')->name('user.edit');
-	Route::patch('users/{id}/edit', 'UserController@update')->name('user.update');
+	Route::get('users', 'UserController@index')->name('user.index')->middleware('can:isAdmin,App\User');
+	Route::delete('users/{id}', 'UserController@destroy')->name('user.destroy')->middleware('can:isAdmin,App\User');;
+	Route::get('users/{id}/edit', 'UserController@edit')->name('user.edit')->middleware('can:isAdmin,App\User');;
+	Route::patch('users/{id}/edit', 'UserController@update')->name('user.update')->middleware('can:isAdmin,App\User');;
